% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_randomization.R
\name{complete_randomization_plus_one_min_one}
\alias{complete_randomization_plus_one_min_one}
\title{Implements the complete randomization design AKA Bernoulli Trial}
\usage{
complete_randomization_plus_one_min_one(n, r)
}
\arguments{
\item{n}{number of observations}

\item{r}{number of randomized designs you would like}
}
\value{
a matrix where each column is one of the \code{r} designs
}
\description{
Implements the complete randomization design AKA Bernoulli Trial
}
\author{
Adam Kapelner
}
