% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frob_norm_sq.R
\name{frob_norm_sq}
\alias{frob_norm_sq}
\title{Naive Frobenius Norm Squared}
\usage{
frob_norm_sq(A)
}
\arguments{
\item{A}{The matrix of interest}
}
\description{
Compute naive / vanilla squared Frobenius Norm of matrix A
}
\author{
Adam Kapelner
}
