% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_and_prints.R
\name{plot.optimal_rerandomization_obj}
\alias{plot.optimal_rerandomization_obj}
\title{Plots a summary of a \code{optimal_rerandomization_obj} object}
\usage{
\method{plot}{optimal_rerandomization_obj}(x, ...)
}
\arguments{
\item{x}{The \code{optimal_rerandomization_obj} object to be summarized in the plot}

\item{...}{The option \code{advanced = TRUE} can be passed here for optimal rerandomization 
results from algorithm type "approx" to see how all the terms in the criterion behave.
You can pass \code{s_min} which controls the minimum number of vectors the plot begins at. 
Below a certain number, the criterion is unstable.
Also, \code{title}, \code{subtitle}, \code{xlab} and \code{ylab} can be passed here.}
}
\description{
Plots a summary of a \code{optimal_rerandomization_obj} object
}
\author{
Adam Kapelner
}
