% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_and_prints.R
\name{plot.W_base_object}
\alias{plot.W_base_object}
\title{Plots a summary of the imbalances in a \code{W_base_object} object}
\usage{
\method{plot}{W_base_object}(x, ...)
}
\arguments{
\item{x}{The \code{W_base_object} object to be summarized in the plot}

\item{...}{\code{title}, \code{subtitle}, \code{xlab}, \code{bins} can 
be specified here to be passed to the ggplot plotting function.
Also \code{log10} can be set to \code{FALSE} to not log the x-axis.}
}
\description{
Plots a summary of the imbalances in a \code{W_base_object} object
}
\author{
Adam Kapelner
}
