% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_rerandomization.R
\name{print.optimal_rerandomization_obj}
\alias{print.optimal_rerandomization_obj}
\title{Prints a summary of a \code{optimal_rerandomization_obj} object}
\usage{
\method{print}{optimal_rerandomization_obj}(x, ...)
}
\arguments{
\item{x}{The \code{optimal_rerandomization_obj} object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a \code{optimal_rerandomization_obj} object
}
\author{
Adam Kapelner
}
