% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_rerandomization.R
\name{print.W_base_object}
\alias{print.W_base_object}
\title{Prints a summary of a \code{W_base_object} object}
\usage{
\method{print}{W_base_object}(x, ...)
}
\arguments{
\item{x}{The \code{W_base_object} object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a \code{W_base_object} object
}
\author{
Adam Kapelner
}
