% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_rerandomization.R
\name{summary.W_base_object}
\alias{summary.W_base_object}
\title{Prints a summary of a \code{W_base_object} object}
\usage{
\method{summary}{W_base_object}(object, ...)
}
\arguments{
\item{object}{The \code{W_base_object} object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a \code{W_base_object} object
}
\author{
Adam Kapelner
}
