% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_rerandomization.R
\name{optimal_rerandomization_tail_approx}
\alias{optimal_rerandomization_tail_approx}
\title{Find the Optimal Rerandomization Design Under the Tail and Kurtosis Approximation}
\usage{
optimal_rerandomization_tail_approx(W_base_object = W_base_object,
  estimator = "linear", q = 0.95, skip_search_length = 1,
  binary_search = FALSE, excess_kurtosis_z = 0, dot_every_x_iters = 100)
}
\arguments{
\item{W_base_object}{An object that contains the assignments to begin with sorted by}

\item{estimator}{"linear" for the covariate-adjusted linear regression estimator (default)
or "difference_in_means".}

\item{q}{The tail criterion's quantile of MSE over z's. The default is 95\%.}

\item{skip_search_length}{In the exhaustive search, how many designs are skipped? Default is 1 for 
full exhaustive search through all assignments provided for in \code{W_base_object}.}

\item{binary_search}{If \code{TRUE}, a binary search is employed to find the optimal threshold instead of 
an exhaustive search. Default is \code{FALSE}.}

\item{excess_kurtosis_z}{An estimate of the excess kurtosis in the measure on z. Default is 0.}

\item{dot_every_x_iters}{Print out a dot every this many iterations. The default is 100. Set to
\code{NULL} for no printout.}
}
\value{
A list containing the optimal design threshold, strategy, and
								other information.
}
\description{
Finds the optimal rerandomization threshold based on a user-defined quantile
and kurtosis based on an approximation of tail standard errors
}
\author{
Adam Kapelner
}
