% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalRerandExpDesigns.R
\docType{package}
\name{OptimalRerandExpDesigns}
\alias{OptimalRerandExpDesigns}
\alias{OptimalRerandExpDesigns-package}
\title{Optimal Rerandomization Threshold Search for Experimental Design}
\description{
A tool to find the optimal rerandomization threshold in non-sequential experiments
}
\references{
Kapelner, A
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{Design,}
\keyword{Experimental}
\keyword{Rerandomization}
