% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_rerandomization.R
\name{plot.optimal_rerandomization_obj}
\alias{plot.optimal_rerandomization_obj}
\title{Plots a summary of a \code{optimal_rerandomization_obj} object}
\usage{
\method{plot}{optimal_rerandomization_obj}(x, ...)
}
\arguments{
\item{x}{The \code{optimal_rerandomization_obj} object to be summarized in the plot}

\item{...}{The option \code{all = TRUE} can be passed here for optimal rerandomization 
results from algorithm type "approx" to see how all the terms in the criterion behave}
}
\description{
Plots a summary of a \code{optimal_rerandomization_obj} object
}
\author{
Adam Kapelner
}
