% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_rerandomization.R
\name{plot.W_base_object}
\alias{plot.W_base_object}
\title{Plots a summary of a \code{W_base_object} object}
\usage{
\method{plot}{W_base_object}(x, ...)
}
\arguments{
\item{x}{The \code{W_base_object} object to be summarized in the plot}

\item{...}{Unused}
}
\description{
Plots a summary of a \code{W_base_object} object
}
\author{
Adam Kapelner
}
